Feat: Utilise google sign-in html element
Utilise google sign-in html element in login page.

Needed to upgrade from google+ sign-in to latest API. Implementation of
hybrid oauth flow is clunkier, so I dropped it in favour of a simpler
server-side verification of the id token. I don't believe the back-end
needs offline access to user data, so I think this is sufficient